Add files via upload
A Python program that stores and processes student result records is in its early stage of development. The student records are stored in a text file called class.txt. In this file, student’s names, student IDs and their marks are saved. Each student occupies one line in the file and each field is separated by tabs. 
A menu is used to list the processing tasks that can be applied on the file. Some of the menu items have been coded; others have a stub (dummy) code provided. The menu program is provided in a Python file called menu.py.